server/pyproject.toml
[tool.poetry]
name = "pandasai-server"
version = "0.1.0"
description = "PandasAI Server"
authors = ["Arslan <arslan@sinaptik.ai>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.9.4"
SQLAlchemy = "^2.0.4"
uvicorn = "^0.20.0"
fastapi = "^0.92.0"
gunicorn = "^20.1.0"
fastapi-event = "^0.1.3"
pydantic = "^1.10.5"
isort = "^5.12.0"
black = "^23.1.0"
pytest = "^7.2.1"
httpx = "^0.23.3"
email-validator = "^1.3.1"
typing-extensions="^4.0.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
python-jose = "^3.3.0"
pylint = "^3.2.2"
pandasai = "^2.1"
pyyaml = "^6.0.1"
passlib = "^1.7.4"
bcrypt = "3.1.7"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
numpy = "1.26.4"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"
pytest-asyncio = "^0.20.3"

[tool.poetry.group.dev.dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
